Check for Potential Cross Account Key Policy Usage (#19)
This PR accounts for #16.

This helper function will check the KMS Key Policy for potential cross account key usage by the following:

AWS Principals with a different account in the ARN.
AWS Principals with a * in the account block of an ARN.
AWS Principals marked as "*"
Principal block as "*"
This is meant to be a MVP and warn Finders Keypers users that there may be cross account usage.

This does not check for:

Condition Keys (yet)
Specific Action (Key Policy) based logic.
Federated, CanonicalUser, and Service Principals
Additionally, it's possible to use a KMS key to encrypt resources in a different account even if the key policy is limited to within the KMS key account. Scenario was tested with S3 and an IAM Principal in the KMS key account calling PutObject. This required a bucket policy that permitted cross-account usage. We consider this an edge case. This PR does not cover this.

For full coverage, it's better to scan all potential accounts and resources. Multi-account coverage is on our roadmap, email us at info@fogsecurity.io if you're interested.